Refactor(tooltipsAnimation):Changing the structure
Changing the way tooltips animation works with IntersectionObserver.
Changing the function of creating tooltips due to having a more structured code and shorter selections.
Changing the placement of the animation and making it modular.
Because the tooltips creation function returns an array, we made a map on the array in the function calling section because each tag only takes elements.

changeLog:[tooltipsAnimation] add to [mapComponent]

Doc(tooltips):Adding various documents and guides to functions related to tooltips